Convert AesPredictor to Predictor
remove aes example file
